Typescriptizing
