+++ .gitignore
@@ -1,0 +1,10 @@
+# dependencies
+node_modules
+
+# production
+dist
+
+# misc
+.DS_Store
+
+npm-debug.log*
