--- content/css/blur.css
@@ -1,21 +1,0 @@
-.blurred {
-  filter: blur(5px);
-}
-
-.mask {
-  border: none;
-  overflow: hidden;
-}
-
-#__inputClone, .mask, .textLayer {
-  position: absolute;
-  top: 0px;
-  left: 0px;
-  border: none;
-  overflow: hidden;
-  white-space: nowrap;
-}
-
-#__inputClone {
-  visibility: hidden;
-}
