+++ public/img/ATTRIBUTIONS.txt
