+++ public/main.html
@@ -4,7 +4,7 @@
 <head>
   <title>Text Blurrer Popup</title>
   <link rel="stylesheet" type="text/css" href="./css/main.css">
-  <script src="./js/main.js" type="module"></script>
+  <script src="./js/bundle.js" type="module"></script>
 </head>
 
 <body>
@@ -53,7 +53,7 @@
     <div class="footer">
       <span id="addUrlsInCurrentTab">+ Add URLs in the current tab</span>
     </div>
-    <textarea title="" id="exclusionInput" placeholder="" rows="10"></textarea>
+    <textarea title="" id="exclusionTextArea" placeholder="" rows="10"></textarea>
   </div>
   <div class="footer">
     <div id="linkContainer">
