+++ src/utils/common.ts
@@ -1,3 +1,3 @@
-export function  escapeRegExp(str) {
+export function  escapeRegExp(str:string):string {
   return str.replace(/([\(\)\{\}\+\*\?\[\]\.\^\$\|\\])/g, '\\$1');
 };
